rewrite sdk and make user usability easier
