+++ avail-go/examples/go-examples/DAdataSubmit/main.go
@@ -1,13 +1,17 @@
 package main
 
 import (
+	// "avail-go-sdk-examples/internal/config"
+
 	"avail-go-sdk/src/config"
 	"avail-go-sdk/src/sdk"
 	"avail-go-sdk/src/sdk/tx"
+
+	// "avail-gsrpc-examples/internal/extrinsics"
+
 	"fmt"
 )
 
-// submitData creates a transaction and makes a Avail data submission
 func main() {
 	config, err := config.LoadConfig()
 	if err != nil {
@@ -17,12 +21,20 @@ func main() {
 	if err != nil {
 		fmt.Printf("cannot create api:%v", err)
 	}
+
+	appID := 0
+
+	// if app id is greater than 0 then it must be created before submitting data
+	if config.AppID != 0 {
+		appID = config.AppID
+	}
+	fmt.Println("Submitting data ...")
 	WaitFor := sdk.BlockInclusion
-	commission := 5
-	BlockHash, txHash, err := tx.Validate(api, config.Seed, WaitFor, commission)
+
+	// submit data
+	BlockHash, txHash, err := tx.SubmitData(api, config.Seed, appID, "my happy data", WaitFor)
 	if err != nil {
 		fmt.Printf("cannot submit data:%v", err)
 	}
 	fmt.Printf("Data submitted successfully with block hash: %v\n and ext hash:%v", BlockHash.Hex(), txHash.Hex())
-
 }
