+++ avail-go/examples/go-examples/balance_AllowDeath/main.go
@@ -4,6 +4,9 @@ import (
 	"avail-go-sdk/src/config"
 	"avail-go-sdk/src/sdk"
 	"avail-go-sdk/src/sdk/tx"
+	"avail-go-sdk/src/sdk/types"
+	"math"
+
 	"fmt"
 )
 
@@ -18,11 +21,13 @@ func main() {
 		fmt.Printf("cannot create api:%v", err)
 	}
 	WaitFor := sdk.BlockInclusion
-	commission := 5
-	BlockHash, txHash, err := tx.Validate(api, config.Seed, WaitFor, commission)
+
+	amount := uint64(math.Pow(10, 18)) * 10 // send amount 10 AVAIL
+	dest := "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty"
+	bondAmountUCompact := types.NewUCompactFromUInt(amount)
+	BlockHash, txHash, err := tx.TransferAllowDeath(api, config.Seed, WaitFor, dest, bondAmountUCompact)
 	if err != nil {
 		fmt.Printf("cannot submit data:%v", err)
 	}
 	fmt.Printf("Data submitted successfully with block hash: %v\n and ext hash:%v", BlockHash.Hex(), txHash.Hex())
-
 }
