+++ avail-go/examples/go-examples/staking_chillOther/main.go
@@ -4,6 +4,7 @@ import (
 	"avail-go-sdk/src/config"
 	"avail-go-sdk/src/sdk"
 	"avail-go-sdk/src/sdk/tx"
+
 	"fmt"
 )
 
@@ -18,11 +19,11 @@ func main() {
 		fmt.Printf("cannot create api:%v", err)
 	}
 	WaitFor := sdk.BlockInclusion
-	commission := 5
-	BlockHash, txHash, err := tx.Validate(api, config.Seed, WaitFor, commission)
+
+	stash := "5GNJqTPyNqANBkUVMN1LPPrxXnFouWXoe2wNSmmEoLctxiZY"
+	BlockHash, txHash, err := tx.ChillOther(api, config.Seed, WaitFor, stash)
 	if err != nil {
 		fmt.Printf("cannot submit data:%v", err)
 	}
 	fmt.Printf("Data submitted successfully with block hash: %v\n and ext hash:%v", BlockHash.Hex(), txHash.Hex())
-
 }
