+++ avail-go/examples/go-examples/staking_nominate/main.go
@@ -4,6 +4,7 @@ import (
 	"avail-go-sdk/src/config"
 	"avail-go-sdk/src/sdk"
 	"avail-go-sdk/src/sdk/tx"
+
 	"fmt"
 )
 
@@ -18,11 +19,11 @@ func main() {
 		fmt.Printf("cannot create api:%v", err)
 	}
 	WaitFor := sdk.BlockInclusion
-	commission := 5
-	BlockHash, txHash, err := tx.Validate(api, config.Seed, WaitFor, commission)
+
+	stash := []string{"5GNJqTPyNqANBkUVMN1LPPrxXnFouWXoe2wNSmmEoLctxiZY", "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty"}
+	BlockHash, txHash, err := tx.Nominate(api, config.Seed, WaitFor, stash)
 	if err != nil {
 		fmt.Printf("cannot submit data:%v", err)
 	}
 	fmt.Printf("Data submitted successfully with block hash: %v\n and ext hash:%v", BlockHash.Hex(), txHash.Hex())
-
 }
