+++ avail-go/examples/go-examples/staking_unbond/main.go
@@ -4,6 +4,9 @@ import (
 	"avail-go-sdk/src/config"
 	"avail-go-sdk/src/sdk"
 	"avail-go-sdk/src/sdk/tx"
+	"avail-go-sdk/src/sdk/types"
+	"math/big"
+
 	"fmt"
 )
 
@@ -18,11 +21,15 @@ func main() {
 		fmt.Printf("cannot create api:%v", err)
 	}
 	WaitFor := sdk.BlockInclusion
-	commission := 5
-	BlockHash, txHash, err := tx.Validate(api, config.Seed, WaitFor, commission)
+
+	bondAmount := new(big.Int)
+	bondAmount.SetString("100000000000000000000", 10)
+
+	// Convert big.Int to types.UCompact
+	bondAmountUCompact := types.NewUCompact(bondAmount)
+	BlockHash, txHash, err := tx.Unbond(api, config.Seed, WaitFor, bondAmountUCompact)
 	if err != nil {
 		fmt.Printf("cannot submit data:%v", err)
 	}
 	fmt.Printf("Data submitted successfully with block hash: %v\n and ext hash:%v", BlockHash.Hex(), txHash.Hex())
-
 }
