+++ avail-go/src/rpc/kate_proof.go
