+++ avail-go/src/rpc/query_proof_data.go
